cleaned up createBasicSchedule. Course instantiation should occur in the controller
