+++ src/main/java/CommandLineInterface.java
@@ -4,7 +4,10 @@ import java.util.List;
 import java.util.Scanner;
 
 public class CommandLineInterface {
-    public static void main(String[] args) {
+
+    public CommandLineInterface() {}
+
+    public ArrayList<String> promptUser() {
         Scanner scanner = new Scanner(System.in);
         Boolean input = false;
         int numCourses = 0;
@@ -24,39 +27,14 @@ public class CommandLineInterface {
             String course = scanner.nextLine();
             courses.add(course);
         }
+        scanner.close();
+        return courses;
+        /**
         Scheduler s = new Scheduler();
         List<Course> instantiatedCourses = CommandLineInterface.courseInstantiator(courses);
         List<Schedule> schedules = s.permutationScheduler(instantiatedCourses);
         //System.out.println(s.createBasicSchedule(courses));
-        scanner.close();
-
-        for(Schedule sch : schedules) {
-            System.out.println(sch);
-        }
+         */
     }
 
-    private static List<Course> courseInstantiator(ArrayList<String> courseCodes) {
-        ArrayList<Course> courses = new ArrayList<>();
-        for (String courseCode : courseCodes) {
-            try {
-                /**
-                 * For every course code, generate the course from CourseCreator, add first lec/tut
-                 * session to the lectures and tutorials within the schedule.
-                 */
-                Course newCourse = CourseCreator.generateCourse(courseCode, 'F');
-                courses.add(newCourse);
-            } catch (IOException exception) {
-                /**
-                 * In case something goes wrong with the API for a specific course code, we print
-                 * the code and the exception that is thrown.
-                 */
-                System.out.println(
-                        "Exception occured for course "
-                                + courseCode
-                                + " with the following message: \n"
-                                + exception.toString());
-            }
-        }
-        return courses;
-    }
 }
